Cleaning up subs from checkpoints
