+++ src/Core/src/Eventuous.Subscriptions/Channels/ChannelWorker.cs
@@ -5,12 +5,12 @@ namespace Eventuous.Subscriptions.Channels;
 public class ChannelWorker<T> {
     readonly Channel<T>              _channel;
     readonly CancellationTokenSource _cts;
-    readonly Task                    _readerTask;
+    readonly Task[]                  _readerTasks;
 
-    public ChannelWorker(Channel<T> channel, Func<T, CancellationToken, ValueTask> process) {
-        _channel    = channel;
-        _cts        = new CancellationTokenSource();
-        _readerTask = Task.Run(() => Read(_cts.Token));
+    public ChannelWorker(Channel<T> channel, Func<T, CancellationToken, ValueTask> process, int concurrencyLevel = 1) {
+        _channel     = channel;
+        _cts         = new CancellationTokenSource();
+        _readerTasks = Enumerable.Range(0, concurrencyLevel).Select(_ => Task.Run(() => Read(_cts.Token))).ToArray();
 
         async Task Read(CancellationToken cancellationToken) {
             try {
@@ -32,7 +32,7 @@ public class ChannelWorker<T> {
         _channel.Writer.Complete();
         _cts.CancelAfter(TimeSpan.FromSeconds(1));
 
-        while (!_readerTask.IsCompleted && !_readerTask.IsCanceled) {
+        while (!_readerTasks.All(x => x.IsCompleted) && !_readerTasks.All(x => x.IsCanceled)) {
             await Task.Delay(10);
         }
