+++ src/Core/test/Eventuous.Tests/TypeRegistrationTests.cs
@@ -5,12 +5,13 @@ namespace Eventuous.Tests;
 public class TypeRegistrationTests {
     readonly TypeMapper _typeMapper;
 
-    public TypeRegistrationTests() => _typeMapper = new TypeMapper();
+    public TypeRegistrationTests() {
+        _typeMapper = new TypeMapper();
+        _typeMapper.RegisterKnownEventTypes(typeof(BookingCancelled).Assembly);
+    }
 
     [Fact]
     public void ShouldResolveDecoratedEvent() {
-        // This test works because event types are registered by the domain Module.cs
-            
         _typeMapper.GetTypeName<BookingCancelled>().Should().Be(TypeNames.BookingCancelled);
         _typeMapper.GetType(TypeNames.BookingCancelled).Should().Be<BookingCancelled>();
     }
